@@ -24,36 +24,33 @@
     "gregwar/captcha": "1.*",
     "guzzlehttp/guzzle": "^7.0",
     "intervention/image": "^2.5",
-    "laravel/framework": "^7.0",
-    "laravel/passport": "^9.2",
-    "lorisleiva/laravel-search-string": "^0.1.6",
+    "laravel/framework": "^8.0",
+    "laravel/passport": "^10.0",
+    "lorisleiva/laravel-search-string": "^1.0",
     "nesbot/carbon": "^2.0",
-    "nunomaduro/collision": "^4.1",
-    "rcrowe/twigbridge": "^0.11.3",
+    "nunomaduro/collision": "^5.0",
+    "rcrowe/twigbridge": "^0.12",
     "spatie/laravel-translation-loader": "^2.6",
     "symfony/process": "^5.0",
     "symfony/yaml": "^5.0",
     "twig/twig": "^2.11",
-    "tymon/jwt-auth": "dev-develop",
+    "tymon/jwt-auth": "^1.0",
     "vectorface/whip": "^0.3.2"
   },
   "require-dev": {
-    "barryvdh/laravel-debugbar": "^3.2",
+    "barryvdh/laravel-debugbar": "^3.5",
     "fzaninotto/faker": "~1.9",
-    "laravel/browser-kit-testing": "^6.0",
-    "laravel/tinker": "^2.2",
-    "mockery/mockery": "1.3.*",
+    "laravel/browser-kit-testing": "^6.1",
+    "laravel/tinker": "^2.4",
+    "mockery/mockery": "^1.4",
     "phpunit/phpunit": "^9.4",
     "symfony/css-selector": "^5.0",
     "symfony/dom-crawler": "^5.0"
   },
   "autoload": {
-    "classmap": [
-      "app/Models",
-      "database"
-    ],
     "psr-4": {
-      "App\\": "app/"
+      "App\\": "app/",
+      "Database\\Factories\\": "database/factories/"
     },
     "files": [
       "app/helpers.php"
