@@ -8,11 +8,11 @@ class HandlerTest extends TestCase
 {
     public function testRenderAjaxException()
     {
-        $json = $this->get('/abc', ['Accept' => 'application/json'])->decodeResponseJson();
+        $json = $this->get('/abc', ['Accept' => 'application/json'])->json();
         $this->assertIsString($json['message']);
         $this->assertTrue($json['exception']);
-        $this->assertTrue(collect($json['trace'])->every(function ($trace) {
-            return Str::startsWith($trace['file'], 'app') && is_int($trace['line']);
-        }));
+        $this->assertTrue(collect($json['trace'])->every(
+            fn ($trace) => Str::startsWith($trace['file'], 'app') && is_int($trace['line'])
+        ));
     }
 }
